adding snake + visual touches + button connections

+++ include/RadioManager.h
@@ -9,12 +9,12 @@ class RadioManager {
 private:
     CC1101::Radio radio;
 
-    int setupRadio(double baud, int length);
+    int setupRadio(float baud, int length);
     int countDataBytes(uint8_t *buf, size_t len);
     int countRepeats(uint8_t *buf, size_t len, int seqLen);
     float scoreCaptures(uint8_t captures[][64], int numCaptures);
     std::vector<uint8_t> extractCycle(std::vector<uint8_t>& data);
-    double baudRateFinder();
+    float baudRateFinder();
 
 public:
     RadioManager(uint8_t cs);
