+++ include/RadioSignal.h
@@ -1,18 +1,18 @@
 #pragma once
 #include "Config.h"
-#include "Signal.h"
+#include "Signal_.h"
 #include <Arduino.h>
 #include <vector>
 
 class RadioSignal : public Signal {
     private:
         std::vector<uint8_t> data;
-        double baudRate = 0;
+        float baudRate = 0;
     public:
         std::vector<uint8_t> getData() const { return data; }
         void setData(std::vector<uint8_t> givenData) { data = givenData; }
         
-        void setBaudRate(double rate) { baudRate = rate; }
+        void setBaudRate(float rate) { baudRate = rate; }
         double getBaudRate() const { return baudRate; }
 
         void transmit() const override;
